Implement basic screen rotation functionality
This commit allows users to rotate the screen 180 degrees. 90 degree
rotations are not yet supported as they cannot be performed as easily. A
180 degree rotation can be done with a simple memory remap.

For the SSD1327 driver this is done by setting the GDDRAM configurations
and for the serial driver we just change the order that bytes are read
from the buffer.